feat(frontend): enable React compiler and refactor rooms page
- Enable experimental React compiler for performance optimization
- Add babel-plugin-react-compiler dependency
- Refactor rooms page to use custom hook for better code organization
