+++ front/app/rooms/page.tsx
@@ -1,9 +1,5 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
-import { useAuth } from "@/lib/auth-context";
-import { Room } from "@/lib/rooms";
 import Link from "next/link";
 import Header from "@/components/Header";
 import { Button } from "@/components/ui/button";
@@ -17,46 +13,10 @@ import {
 } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Loader2 } from "lucide-react";
+import { useRooms } from "./hooks";
 
 export default function RoomsPage() {
-  const { token, logout } = useAuth();
-  const [rooms, setRooms] = useState<Room[]>([]);
-  const [loading, setLoading] = useState(true);
-  const router = useRouter();
-  useEffect(() => {
-    const fetchRooms = async () => {
-      if (!token) {
-        return;
-      }
-
-      try {
-        setLoading(true);
-        const roomsResponse = await fetch(`/api/rooms`, {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-
-        if (roomsResponse.status === 401) {
-          logout();
-          return;
-        }
-
-        const roomsData = await roomsResponse.json();
-        const newToken = roomsResponse.headers.get("X-New-Token");
-        if (newToken) {
-          localStorage.setItem("token", newToken);
-        }
-        setRooms(roomsData);
-      } catch (error) {
-        console.error("Error fetching rooms:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchRooms();
-  }, [token, logout]);
+  const { loading, rooms, router } = useRooms();
 
   return (
     <div className="min-h-screen bg-gray-50 dark:bg-gray-900">
