+++ front/package.json
@@ -15,6 +15,7 @@
     "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.2",
     "axios": "^1.8.2",
+    "babel-plugin-react-compiler": "19.0.0-beta-3229e95-20250315",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.479.0",
